Update dockyards release controller to reconcile image factory urls
Currently the talos installer releases only reconcile the latest version
of the release ranges.

This updates the dockyard release controller to reconcile the talos
image factory url for the version and platform. If not annotated with a
specific schematic the release uses the default schematic without any
customization.

DY-431 #done.